# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NOAA searcher.",
    "version": "1.0.0",
    "description": "This searches for NOAA records by station id.  Courtesy of Erick Stone."
  },
  "basePath": "/",
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "search",
      "description": "Search stuff"
    }
  ],
  "paths": {
    "/search": {
      "get": {
        "tags": ["search"],
        "produces": ["application/json"],
        "parameters": [{
          "in": "query",
          "name": "station_id",
          "description": "Id of the station.",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "Returns a list of stations matching criteria using uri params style.",
            "headers":  {},
            "examples": {
              "station_id":"AE000041196",
              "year_month":"20170101",
              "element":"TAVG",
              "value1":"217",
              "mflag1":"H",
              "qflag1":"",
              "sflag1":"S",
              "value2":"",
              "MFLAG2":null,
              "QFLAG2":null,
              "SFLAG2":null
            }
          }
        }
      }
    },
    "/search/:station_id": {
      "get": {
        "produces": ["application/json"],
        "parameters": [{
          "in": "path",
          "name": "station_id",
          "description": "Id of the station.",
          "type": "string",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "Returns a list of stations matching criteria using uri params style.",
            "headers":  {},
            "examples": {
              "station_id":"AE000041196",
              "year_month":"20170101",
              "element":"TAVG",
              "value1":"217",
              "mflag1":"H",
              "qflag1":"",
              "sflag1":"S",
              "value2":"",
              "MFLAG2":null,
              "QFLAG2":null,
              "SFLAG2":null
            }
          }
        }
      }
    }
  }
}